requirements.txt
google-generativeai>=0.5.4
newsapi-python>=0.2.7
python-dotenv>=1.0.0
pydantic>=2.0
requests>=2.0
google-api-python-client>=2.172.0
google-auth>=2.0
pydantic-settings>=2.0
google-auth-oauthlib>=1.0
pre-commit>=4.0
pyyaml>=6.0